Move Bukkit API calls to separate services
